conflicts solved
