+++ src/routes/entries.js
@@ -4,4 +4,6 @@ const entriesController = require('../controllers/entries-controller');
 
 router.get("/news", entriesController.getNewsEntries);
 
+router.put("/news/:id" ,entriesController.updateNewsEntry);
+
 module.exports = router;
